Move hardhat/example.config.ts to example-project
