+++ v-next/hardhat/README.md
@@ -1,3 +1,3 @@
 # Hardhat `v-next`
 
-`node --import tsx/esm src/cli.ts --config src/example.config.ts example`
+To try this out see `../example-project`
