+++ v-next/hardhat/test/internal/cli/helpers/getGlobalHelpString.ts
@@ -3,7 +3,7 @@ import type { Task } from "@ignored/hardhat-vnext-core/types/tasks";
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 
-import { readClosestPackageJson } from "@nomicfoundation/hardhat-utils/package";
+import { readClosestPackageJson } from "@ignored/hardhat-vnext-utils/package";
 
 import { getGlobalHelpString } from "../../../../src/internal/cli/helpers/getGlobalHelpString.js";
