+++ v-next/hardhat/test/internal/cli/main.ts
@@ -25,7 +25,6 @@ import { HardhatError } from "@ignored/hardhat-vnext-errors";
 import { isCi } from "@ignored/hardhat-vnext-utils/ci";
 import chalk from "chalk";
 
-import packageJson from "../../../package.json";
 import {
   main,
   parseGlobalOptions,
@@ -225,7 +224,7 @@ describe("main", function () {
           lines = msg;
         });
 
-        const expected = `Hardhat version ${packageJson.version}
+        const expected = `Hardhat version ${await getHardhatVersion()}
 
 Usage: hardhat [GLOBAL OPTIONS] <TASK> [SUBTASK] [TASK OPTIONS] [--] [TASK ARGUMENTS]
